feat(docs-pages): lift /docs routes into shared package
Introduces @getmunin/docs-pages so munin-cloud can mount the same
developer portal under its own auth/header chrome instead of bouncing
visitors to a separate marketing site.

- 28 docs files moved from apps/web/app/[locale]/docs/ into the new
  package; apps/web routes collapse to 10 one-liner re-export shells.
- The package owns its i18n routing (./i18n-navigation) so it has no
  app-local aliases.
- @getmunin/backend-core now publishes the OpenAPI spec and docs
  fixtures via subpath exports (@getmunin/backend-core/openapi.json,
  @getmunin/backend-core/docs-fixtures/*) and includes them in files
  so downstream consumers can import them at SSR/build time.

Also drops the redundant CONV pill from the Last conversations rows,
and narrows the brace-expansion override to the affected v5 range so
older minimatch consumers keep working.
