+++ THIRD_PARTY_LICENSES.md
@@ -2,7 +2,7 @@
 
 This file lists open-source dependencies bundled with Munin runtime artifacts, together with their license notices. It is generated by `pnpm licenses:generate` from `pnpm licenses list --prod --json` and verified in CI.
 
-Generated against 845 production dependencies.
+Generated against 846 production dependencies.
 
 ---
 
@@ -17078,7 +17078,7 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 ---
 
-## balanced-match@4.0.4
+## balanced-match@1.0.2, 4.0.4
 
 > Match balanced character pairs, like "{" and "}"
 - Homepage: https://github.com/juliangruber/balanced-match#readme
@@ -17087,9 +17087,7 @@ SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 ```
 (MIT)
 
-Original code Copyright Julian Gruber <julian@juliangruber.com>
-
-Port to TypeScript Copyright Isaac Z. Schlueter <i@izs.me>
+Copyright (c) 2013 Julian Gruber &lt;julian@juliangruber.com&gt;
 
 Permission is hereby granted, free of charge, to any person obtaining a copy of
 this software and associated documentation files (the "Software"), to deal in
@@ -17482,7 +17480,7 @@ Original Author, when distributed with the Software.
 
 ---
 
-## brace-expansion@5.0.6
+## brace-expansion@1.1.14, 2.1.0, 5.0.6
 
 > Brace expansion as known from sh/bash
 - Homepage: https://github.com/juliangruber/brace-expansion#readme
@@ -17491,9 +17489,7 @@ Original Author, when distributed with the Software.
 ```
 MIT License
 
-Copyright Julian Gruber <julian@juliangruber.com>
-
-TypeScript port Copyright Isaac Z. Schlueter <i@izs.me>
+Copyright (c) 2013 Julian Gruber <julian@juliangruber.com>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -19078,6 +19074,36 @@ IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR
 ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, 
 TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE 
 SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+```
+
+---
+
+## concat-map@0.0.1
+
+> concatenative mapdashery
+- Homepage: https://github.com/substack/node-concat-map#readme
+- Author: James Halliday
+- License: MIT
+
+```
+This software is released under the MIT license:
+
+Permission is hereby granted, free of charge, to any person obtaining a copy of
+this software and associated documentation files (the "Software"), to deal in
+the Software without restriction, including without limitation the rights to
+use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+the Software, and to permit persons to whom the Software is furnished to do so,
+subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 ```
 
 ---
