+++ package.json
@@ -51,7 +51,7 @@
       "fast-uri": "^3.1.2",
       "kysely": "^0.28.17",
       "ws": "^8.20.1",
-      "brace-expansion": "^5.0.6"
+      "brace-expansion@>=5.0.0 <5.0.6": "^5.0.6"
     }
   }
 }
