+++ packages/docs-pages/eslint.config.mjs
@@ -1,0 +1,10 @@
+import tseslint from 'typescript-eslint';
+import shared from '@getmunin/eslint-config';
+
+export default [
+  ...shared,
+  {
+    files: ['eslint.config.mjs'],
+    ...tseslint.configs.disableTypeChecked,
+  },
+];
