+++ packages/docs-pages/package.json
@@ -1,0 +1,49 @@
+{
+  "name": "@getmunin/docs-pages",
+  "version": "0.1.0",
+  "license": "MIT",
+  "description": "Munin developer-portal route components shared between OSS and cloud Next apps.",
+  "publishConfig": {
+    "registry": "https://npm.pkg.github.com",
+    "access": "restricted"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/getmunin/munin.git",
+    "directory": "packages/docs-pages"
+  },
+  "type": "module",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./docs.css": "./src/docs.css"
+  },
+  "files": [
+    "src"
+  ],
+  "scripts": {
+    "typecheck": "tsc -p tsconfig.json --noEmit",
+    "lint": "eslint src"
+  },
+  "dependencies": {
+    "@getmunin/backend-core": "workspace:*"
+  },
+  "peerDependencies": {
+    "next": "^15.0.0",
+    "next-intl": "^4.0.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-markdown": "^9.0.0"
+  },
+  "devDependencies": {
+    "@getmunin/eslint-config": "workspace:*",
+    "@getmunin/tsconfig": "workspace:*",
+    "@types/react": "^19.0.0",
+    "@types/react-dom": "^19.0.0",
+    "eslint": "^9.10.0",
+    "next": "^15.5.18",
+    "react-markdown": "^9.1.0",
+    "typescript": "^5.6.0"
+  }
+}
