+++ packages/docs-pages/src/_components/auth-chip.tsx
