+++ packages/docs-pages/src/_components/mcp-card.tsx
