+++ packages/docs-pages/src/_components/mcp-sidebar.tsx
