+++ packages/docs-pages/src/_lib/search-index.ts
