+++ packages/docs-pages/src/i18n-navigation.ts
@@ -1,0 +1,11 @@
+import { createNavigation } from 'next-intl/navigation';
+import { defineRouting } from 'next-intl/routing';
+
+const routing = defineRouting({
+  locales: ['en', 'nb'],
+  defaultLocale: 'en',
+  localePrefix: 'always',
+});
+
+export const { Link, redirect, usePathname, useRouter, getPathname } =
+  createNavigation(routing);
