+++ packages/docs-pages/src/skills/[...slug]/page.tsx
@@ -1,0 +1,35 @@
+import { notFound } from 'next/navigation';
+import { Link } from '../../i18n-navigation';
+import ReactMarkdown from 'react-markdown';
+import { findSkill, groupByModule, skills } from '../../_lib/skills';
+import { SkillsSidebar } from '../../_components/skills-sidebar';
+
+export function generateStaticParams() {
+  return skills.map((s) => ({ slug: [s.module, s.slug] }));
+}
+
+export default async function SkillDetail({ params }: { params: Promise<{ slug: string[] }> }) {
+  const { slug } = await params;
+  const [moduleName, slugName] = slug;
+  if (!moduleName || !slugName) notFound();
+  const skill = findSkill(moduleName, slugName);
+  if (!skill) notFound();
+  const groups = groupByModule();
+  return (
+    <>
+      <SkillsSidebar groups={groups} />
+      <main className="docs-main">
+        <div className="skill-detail">
+          <div className="breadcrumb">
+            <Link href="/docs/skills">← Back to skills</Link>
+            <span className="uri">skill://{skill.module}/{skill.slug}</span>
+            <span className="title-pin">{skill.title}</span>
+          </div>
+          <div className="markdown">
+            <ReactMarkdown>{skill.content}</ReactMarkdown>
+          </div>
+        </div>
+      </main>
+    </>
+  );
+}
