+++ packages/docs-pages/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@getmunin/tsconfig/react-library.json",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "resolveJsonModule": true
+  },
+  "include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.json"],
+  "exclude": ["node_modules"]
+}
